handle 'exit' message
At least with the helix editor, prior to this change,
golangci-lint-langserver will handle the 'shutdown' message,
but will then stall, waiting for the invoking client (helix)
to terminate the LSP process, while helix is waiting for the LSP process
to self-terminate upon receiving the 'exit' message.
Eventually, helix will terminate the LSP process
after a ~10 second grace period has passed,
though this is rather annoying to wait for.

With the change, the 'exit' command will cause
the jsonrpc2 connection to close,
which in turn will lead to the LSP process exiting quickly.